# swagger.yaml
openapi : 3.0.0

info:
     title: Learn Express and Swagger -LearnCodeOnline
     description: LCO - a course Segment About write docs
     version: 1.0.0
     contact:
       email : sundar@google.com
       url : "https://github.com/Sundarleo11"
     
servers: 
- url: "{protocol}://localhost:5000/api/{version}"
  description: Development server for local Host
  variables:
   version:
    enum:
     - v1
     - v2
    default: v1
   protocol:
    enum:
    - http
    - https
    default: http

components:
 securitySchemes:

   ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

   BearerAuth:
      type: http
      scheme: bearer 


paths:
  /lco:
    get:
     tags:
      - string
     summary: return a greet message from lco
     responses: 
      200:
        description: All good success
        content:
          application/json:
           schema:
            type: string
            example : "mystring"

      400:
        description: Bad request
      500:      
        description: internal server error
  /lcoObject:
    get:
     tags:  
      - Object
     summary: return unique course
     responses: 
        200:
         description: All good success
         content:
            application/json:
               schema:
                 type: object
                 properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: string  

        400:
         description: Bad request
        500:      
         description: internal server error   

  /course:
    get:
     tags:  
      - Array
     summary: return all courses
     responses: 
        200:
         description: All good success
         content:
             application/json:
                schema:
                  type: array
                  items:
                   type: object
                   properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: string

        400:
         description: Bad request
        500:      
         description: internal server error
     